app/requirements.txt
Flask
tensorflow
numpy
Pillow
opencv-python
requests
flask-cors
